+++ Mods ChoGGi/Fix Dust Devils Block Building/metadata.lua
@@ -1,0 +1,22 @@
+return PlaceObj("ModDef", {
+	"title", "Fix Dust Devils Block Building",
+	"id", "ChoGGi_FixDustDevilsBlockBuilding",
+	"steam_id", "2607568970",
+	"pops_any_uuid", "cb4f6bc8-cdb8-4791-91a0-1b508f5aee09",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+No more cheesing dust devils with waste rock depots (etc).
+
+Somewhat requested by Visolaris.
+]],
+})