+++ Mods ChoGGi/Solar Array Follows Sun/metadata.lua
@@ -12,9 +12,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "1570918489",
 	"pops_any_uuid", "105b752a-e526-4055-9009-c253bbe0e3fa",
 	"lua_revision", 1007000, -- Picard
-	"version", 6,
+	"version", 7,
 	"version_major", 0,
-	"version_minor", 6,
+	"version_minor", 7,
 	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {